Update Classes Service / Repository

+++ src/main/java/com/example/ceresto/eat/repository/DiningTableRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiningTableRepository extends JpaRepository<DiningTable, Integer> {
+public interface DiningTableRepository extends JpaRepository<DiningTable, Long> {
 }
